test(executor): split per-call streaming and stateful result assembly tests into request file
